refactor: get_judge_result 함수의 위치 변경 및 래퍼 함수 추가
