+++ core/controller.py
@@ -154,25 +154,6 @@ def get_case_summary():
     chain = prompt | llm | StrOutputParser()
     return chain.invoke({})
 
-def get_judge_result(message_list):
-    llm = get_llm()
-    prompt = ChatPromptTemplate.from_template("""
-        당신은 공정한 AI 판사입니다. 아래는 검사와 변호사의 주장 및 증거입니다. 
-        이 정보를 바탕으로 누가 더 논리적이고 설득력 있는 주장을 했는지 판단해주세요.
-        그 이유도 간단히 설명해주세요.
-
-        [전체 대화 내용]:
-        {messages}
-
-        판결 형식:
-        [승자]: (검사 또는 변호사)
-        [판결 이유]: (간단하고 논리적인 이유)
-    """)
-
-    messages_joined = "\n".join([f"[{m['role']}]: {m['content']}" for m in message_list if m['role'] in ['검사', '변호사']])
-    chain = prompt | llm | StrOutputParser()
-    return chain.invoke({"messages": messages_joined})
-
 
 def make_case_judgment_prompt(case: dict) -> str:
     return f"""
@@ -277,3 +258,14 @@ def ask_witness_wrapper(question, name, wtype, case_summary):
 def ask_defendant_wrapper(question, defendant_name, case_summary):
     from interrogation.interrogator import ask_defendant
     return ask_defendant(question, defendant_name, case_summary)
+
+
+#==============================================
+# verdict.py의 함수
+# 단순히 함수 호출 목적
+# verdict.py -> chat.py로 넘겨줌
+#==============================================  
+
+def get_judge_result_wrapper(message_list):
+    from verdict import get_judge_result
+    return get_judge_result(message_list)
